+++ hirpc-kit/CHANGELOG.md
@@ -2,6 +2,12 @@
 
 All notable changes to this package will be documented in this file.
 
+## [0.1.3] - 2025-04-12
+
+### Changed
+
+- Refactor imports to check for proxy domain.
+
 ## [0.1.2] - 2025-03-22 
 
 ### Added